requirements.txt
python-telegram-bot==13.14
revChatGPT==0.0.14
python-dotenv==0.21.0